finished

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: apereira <apereira@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/23 17:06:20 by apereira          #+#    #+#             */
-/*   Updated: 2022/11/25 17:18:47 by apereira         ###   ########.fr       */
+/*   Updated: 2022/11/28 14:19:07 by apereira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,12 +19,12 @@
 
 int	ft_putstr(char *s);
 int	ft_putchar(int c);
-int	ft_to_hex(int address, char c);
+int	ft_to_hex(unsigned int address, char c);
 int	ft_check(va_list arg, char type);
 int	ft_printf(const char *string, ...);
 int	ft_putnbr(int n, int count);
 int	ft_putunbr(unsigned int n, int count);
-int	ft_to_uhex(unsigned long address, int count);
+int	ft_to_uhex(unsigned long long address, int count);
 int	ft_nbr_len(unsigned long number, int div_by);
 
 #endif
